compiler crashing
