+++ src/compiler/parser/mod.rs
@@ -1,5 +1,7 @@
-mod lines_0_1000;
-pub use lines_0_1000::*;
+mod lines_0_500;
+pub use lines_0_500::*;
+mod lines_500_1000;
+pub use lines_500_1000::*;
 mod lines_1000_2000;
 pub use lines_1000_2000::*;
 mod lines_2000_3000;
